Refactored prompts to an individual file
